pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "doc-proofreader"
version = "0.4.0"
description = "A powerful document proofreading tool with multi-model LLM support including Gemini 2.5 Pro, Claude, and GPT via OpenRouter."
authors = [{name = "caerulex", email = "timelesswind@gmail.com"}]
readme = "README.md"
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "python-docx>=1.1.0",
    "openai>=1.10.0",
    "python-dotenv>=1.0.1"
]

[project.optional-dependencies]
dev = [
    "black",
    "ruff",
]
test = [
    "pytest",
]

[project.urls]
Homepage = "http://example.com/"
Repository = "http://example.com/"

[tool.setuptools]
py-modules = []
